requirements.txt
streamlit
openai-whisper
transformers
torch
torchaudio